Prevent URI Corruption
In some cases (retry? multiple poll?) we saw the Input URI value become corrupted when service discovery would happen more than once.

The original URI would end up prefixed multiple times with SD http://host:port values.

This change uses the original URI value if it begins with a / character, otherwise tries to parse it as a URL, extracts the path value, and then uses the SD URL + path value as the URI for the request.